chatbotapi/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
google-generativeai==0.5.0
